90
90

+++ native/src/main/jni/animation/thorvg/inc/thorvg.h
@@ -508,7 +508,7 @@ public:
      *
      * @since 1.0
      */
-    uint8_t ref() noexcept;
+    uint16_t ref() noexcept;
 
     /**
      * @brief Decrement the reference count for the Paint instance.
@@ -525,7 +525,7 @@ public:
      *
      * @since 1.0
      */
-    uint8_t unref(bool free = true) noexcept;
+    uint16_t unref(bool free = true) noexcept;
 
     /**
      * @brief Retrieve the current reference count of the Paint instance.
@@ -539,7 +539,7 @@ public:
      *
      * @since 1.0
      */
-    uint8_t refCnt() const noexcept;
+    uint16_t refCnt() const noexcept;
 
     /**
      * @brief Returns the ID value of this class.
@@ -888,6 +888,8 @@ public:
      * @retval Result::InvalidArguments in case the radius @p r or @p fr value is negative.
      *
      * @note In case the radius @p r is zero, an object is filled with a single color using the last color specified in the colorStops().
+     * @note In case the focal point (@p fx and @p fy) lies outside the end circle, it is projected onto the edge of the end circle.
+     * @note If the start circle doesn't fully fit inside the end circle (after possible repositioning), the @p fr is reduced accordingly.
      * @note By manipulating the position and size of the focal point, a wide range of visual effects can be achieved, such as directing
      * the gradient focus towards a specific edge or enhancing the depth and complexity of shading patterns.
      * If a focal effect is not desired, simply align the focal point (@p fx and @p fy) with the center of the end circle (@p cx and @p cy)
@@ -1095,14 +1097,17 @@ public:
     /**
      * @brief Sets the dash pattern of the stroke.
      *
-     * @param[in] dashPattern The array of consecutive pair values of the dash length and the gap length.
+     * @param[in] dashPattern An array of alternating dash and gap lengths.
      * @param[in] cnt The length of the @p dashPattern array.
-     * @param[in] offset The shift of the starting point within the repeating dash pattern from which the path's dashing begins.
+     * @param[in] offset The shift of the starting point within the repeating dash pattern, from which the pattern begins to be applied.
      *
-     * @retval Result::InvalidArguments In case @p dashPattern is @c nullptr and @p cnt > 0, @p cnt is zero, any of the dash pattern values is zero or less.
+     * @retval Result::InvalidArguments In case @p dashPattern is @c nullptr and @p cnt > 0 or @p dashPattern is not @c nullptr and @p cnt is zero.
      *
      * @note To reset the stroke dash pattern, pass @c nullptr to @p dashPattern and zero to @p cnt.
-     * @warning @p cnt must be greater than 1 if the dash pattern is valid.
+     * @note Values of @p dashPattern less than zero are treated as zero.
+     * @note If all values in the @p dashPattern are equal to or less than 0, the dash is ignored.
+     * @note If the @p dashPattern contains an odd number of elements, the sequence is repeated in the same
+     * order to form an even-length pattern, preserving the alternation of dashes and gaps.
      *
      * @since 1.0
      */
